AsaphisToolBackend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
rembg[gpu]>=2.0.50
